include typescript, light & dark mode
